+++ frontend/.next/server/server-reference-manifest.js
@@ -1,1 +1,1 @@
-self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"sCUVtIVFnjYUYxKh91gbgvMniOrB9xknsIVPFgv7njg=\"\n}"
+self.__RSC_SERVER_MANIFEST="{\n  \"node\": {},\n  \"edge\": {},\n  \"encryptionKey\": \"rJPKSfLXu6H8MVpDi7d0FpvlXvz6CB8n6wvHoohjxmQ=\"\n}"