+++ frontend/.next/server/webpack-runtime.js
@@ -130,7 +130,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	(() => {
-/******/ 		__webpack_require__.h = () => ("7de162d8d76e6d8e")
+/******/ 		__webpack_require__.h = () => ("60f09c3902b5d796")
 /******/ 	})();
 /******/ 	
 /******/ 	/* webpack/runtime/hasOwnProperty shorthand */