+++ frontend/.next/types/app/page.ts
@@ -1,4 +1,4 @@
-// File: /mnt/c/Users/jizar/OneDrive/Documents/DiscordBot/frontend/app/page.tsx
+// File: C:\Users\jizar\OneDrive\Documents\DiscordBot\frontend\app\page.tsx
 import * as entry from '../../../app/page.js'
 import type { ResolvingMetadata, ResolvingViewport } from 'next/dist/lib/metadata/types/metadata-interface.js'